requirements.txt
telethon==1.40.0
python-dotenv==1.1.0
mcp>=1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
